refactor client folder
