+++ client/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { SignupComponent } from './signup/signup.component';
-import {InventoryComponent} from './inventory/inventory.component';
+import { HomeComponent } from './pages/home/home.component';
+import { SignupComponent } from './pages/signup/signup.component';
+import {InventoryComponent} from './pages/inventory/inventory.component';
 
 const routes: Routes = [
   {path: '', component: HomeComponent},
